SonicSensorの値をdoubleに変更する

+++ lib/minimum/sonic_sensor.h
@@ -7,7 +7,7 @@ class SonicSensor {
 public:
     virtual void Update() = 0;
 
-    virtual float GetRawValue() = 0;
+    virtual double GetRawValue() = 0;
 };
 
 #endif //LABORATORY_WORK_IN_INFORMATION_ENGINEERING_III_SONIC_SENSOR_H
